feat: Add option to save temporary .go file
The commit adds functionality to save the temporary .go file generated during the creation of the bundle executable. The user is prompted to choose whether they want to save the file or not. If they choose to save, the file is renamed with a .saved extension. If they choose not to save, the file is deleted.